FINALLY FIXED THIS FUCKIN SAVE SYSTEM
god that took so much out of me
singular set-back and i was considering abandoning this whole project. that took so much longer to fix then i thought it would
it was only like a day or two, but i usually fix this kind of issue within the first instance of it appearing. its so rare for something to take me this long to fix.
guess im also just not really used to fixing code. im more used to just writing new code and re-using old code. rarely gotta fix old code though